+++ src/main/java/com/example/api_plantanciones/dto/sensor/SensorHumTemAvgHist.java
@@ -1,4 +1,4 @@
 package com.example.api_plantanciones.dto.sensor;
 
-public record SensorHumTemAvgHist(Long idSensor, Long avgTemp, long avgHum) {
+public record SensorHumTemAvgHist(Long idSensor, Double avgTemp, Double avgHum) {
 }
